apps/face-detection/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "face_detection" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../coverage/apps/face-detection/html' --cov-report xml:'../../coverage/apps/face-detection/coverage.xml' --html='../../reports/apps/face-detection/unittests/html/index.html' --junitxml='../../reports/apps/face-detection/unittests/junit.xml'"

[tool.poetry]
name = "face-detection"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "face_detection"

  [tool.poetry.dependencies]
  python = ">=3.9,<3.11"
  deepface = "^0.0.85"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.0.2"
  flake8 = "6.0.0"
  pytest = "7.3.1"
  pytest-sugar = "0.9.7"
  pytest-cov = "4.1.0"
  pytest-html = "3.2.0"
  grpcio = "^1.62.0"
  grpcio-tools = "^1.62.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
